tools/omnissm - fix config event unmarshaling issue
ConfigurationItem events already needed special handling because the
"state" field could be either null, a string, or an object containing
fields "code" and "name". However, the object's "code" field could also
be present in the event payload as a float (i.e. "16.0" rather than
"16") - this caused an unclear error message to be generated during
function startup when the event body was unmarshaled. Because the "code"
field is not currently being used, it has been removed from the
intermediary struct type.